+++ lamb/service/redis/throttling.py
@@ -18,11 +18,12 @@ from lamb.utils import dpath_value
 
 import redis
 
-__all__ = ["redis_rate_check_pipelined", "redis_rate_check_lua", "redis_rate_clear_lua", "RateLimit"]
-
 logger = logging.getLogger(__name__)
 
+__all__ = ["redis_rate_check_pipelined", "redis_rate_check_lua", "redis_rate_clear_lua"]
 
+
+# utilities
 def _check_limits(limits: List[Tuple[int, int]]) -> Dict[int, int]:
     limits_dict: Dict[int, int] = {}
     for limit, duration in limits:
@@ -41,15 +42,77 @@ def _check_limits(limits: List[Tuple[int, int]]) -> Dict[int, int]:
 
 
 def _hash_marked_bucket_name(value: str) -> str:
-    """Add hash_tag mark to bucket name of not exist already"""
+    """Add hash_tag mark to bucket name if not exist already"""
     if not re.match(r"^(.*?)\{(.+?)\}(.*?)$", value):
         return f"{{{value}}}"
     else:
         return value
 
 
-def redis_rate_check_pipelined(conn: redis.Redis, bucket_name_base: str, limits: List[Tuple[int, int]] = list()):
+@dataclasses.dataclass(frozen=True)
+class _RateLimit:
+    key: str
+    limit: int
+    current: int
+    duration: int
+    time_slot: int
+
+    @property
+    def success(self) -> bool:
+        return self.current <= self.limit
+
+
+def _redis_rate_parse_limits(bucket_name_base, limits: List[Tuple[int, int]]) -> List[_RateLimit]:
     # check and sort limits
+    limits_dict: Dict[int, int] = _check_limits(limits)
+    bucket_name_base = _hash_marked_bucket_name(bucket_name_base)
+
+    # repack bucket and limits
+    now = int(time.time())
+
+    result = []
+    for duration, limit in limits_dict.items():
+        time_slot = now // duration
+        bucket_name_slot = f"{bucket_name_base}:{duration}:{time_slot}"
+
+        rate_limit = _RateLimit(key=bucket_name_slot, limit=limit, current=0, duration=duration, time_slot=time_slot)
+        result.append(rate_limit)
+
+    return result
+
+
+def _redis_rate_parse_response(response: Dict[str, dict]) -> List[_RateLimit]:
+    """Internal function to parse response from Redis API"""
+    result = []
+    for key, res in response.items():
+        subkey, _, time_slot = key.rpartition(":")
+        time_slot = int(time_slot)
+
+        subkey, _, duration = subkey.rpartition(":")
+        duration = int(duration)
+
+        current = dpath_value(res, "current", int)
+        limit = dpath_value(res, "limit", int)
+
+        rate_limit = _RateLimit(key=key, limit=limit, current=current, duration=duration, time_slot=time_slot)
+        result.append(rate_limit)
+
+    return result
+
+
+# throttling services
+def redis_rate_check_pipelined(conn: redis.Redis, bucket_name_base: str, limits=None):
+    """Throttling service based on Redis pipelines
+
+    :param conn: connection to Redis instance
+    :param bucket_name_base: bucket name prefix that would be used for storing throttling data
+    :param limits: List ot throttling limits as (limit, seconds)
+
+    :raise ThrottlingError: in case of throttling limits reach
+    """
+    # check and sort limits
+    if limits is None:
+        limits = list()
     logger.debug(f"start rate check pipelined: {conn, bucket_name_base, limits}")
     limits_dict: Dict[int, int] = _check_limits(limits)
 
@@ -69,61 +132,24 @@ def redis_rate_check_pipelined(conn: redis.Redis, bucket_name_base: str, limits:
             raise ThrottlingError
 
 
-# lua version
-@dataclasses.dataclass(frozen=True)
-class RateLimit:
-    key: str
-    limit: int
-    current: int
-    duration: int
-    time_slot: int
-
-    @property
-    def success(self) -> bool:
-        return self.current <= self.limit
-
-
-def _redis_rate_parse_limits(bucket_name_base, limits: List[Tuple[int, int]]) -> List[RateLimit]:
-    # check and sort limits
-    limits_dict: Dict[int, int] = _check_limits(limits)
-    bucket_name_base = _hash_marked_bucket_name(bucket_name_base)
-
-    # repack bucket and limits
-    now = int(time.time())
-
-    result = []
-    for duration, limit in limits_dict.items():
-        time_slot = now // duration
-        bucket_name_slot = f"{bucket_name_base}:{duration}:{time_slot}"
-
-        rate_limit = RateLimit(key=bucket_name_slot, limit=limit, current=0, duration=duration, time_slot=time_slot)
-        result.append(rate_limit)
-
-    return result
-
-
-def _redis_rate_parse_response(response: Dict[str, dict]) -> List[RateLimit]:
-    result = []
-    for key, res in response.items():
-        subkey, _, time_slot = key.rpartition(":")
-        time_slot = int(time_slot)
-
-        subkey, _, duration = subkey.rpartition(":")
-        duration = int(duration)
-
-        current = dpath_value(res, "current", int)
-        limit = dpath_value(res, "limit", int)
-
-        rate_limit = RateLimit(key=key, limit=limit, current=current, duration=duration, time_slot=time_slot)
-        result.append(rate_limit)
-
-    return result
-
-
 def redis_rate_check_lua(
-    conn: redis.Redis, bucket_name_base: str, limits: List[Tuple[int, int]] = list(), increment: bool = True
-) -> List[RateLimit]:
+    conn: redis.Redis,
+    bucket_name_base: str,
+    limits=None,
+    increment: bool = True,
+) -> List[_RateLimit]:
+    """Throttling service based on Redis LUA scripting over throttling buckets
+
+    :param conn: connection to Redis instance
+    :param bucket_name_base: bucket name prefix that would be used for storing throttling data
+    :param limits: List ot throttling limits as (limit, seconds)
+    :param increment: Use incremental logic
+
+    :raise ThrottlingError: in case of throttling limits reach
+    """
     # check and sort limits
+    if limits is None:
+        limits = list()
     logger.debug(f"Lua throttling. Start rate check: {conn, bucket_name_base, limits, increment}")
     rate_limits = _redis_rate_parse_limits(bucket_name_base, limits)
     logger.debug(f"Lua throttling. Limits: {rate_limits}")
@@ -204,10 +230,19 @@ def redis_rate_check_lua(
 
 
 def redis_rate_clear_lua(
-    conn: redis.Redis, bucket_name_base: str, limits: List[Tuple[int, int]] = list(), increment: bool = True
+    conn: redis.Redis,
+    bucket_name_base: str,
+    limits=None,
 ):
+    """Utility function to clean LUA throttling buckets
+    :param conn: connection to Redis instance
+    :param bucket_name_base: bucket name prefix that would be used for storing throttling data
+    :param limits: List ot throttling limits as (limit, seconds)
+    """
     # check and sort limits
-    logger.debug(f"Lua throttling. Start rate clear: {conn, bucket_name_base, limits, increment}")
+    if limits is None:
+        limits = list()
+    logger.debug(f"Lua throttling. Start rate clear: {conn, bucket_name_base, limits}")
     rate_limits = _redis_rate_parse_limits(bucket_name_base, limits)
     logger.info(f"Lua throttling. Limits to clear: {rate_limits}")